Data transfer objects added, controllers and services updated

+++ src/main/java/com/borovyknt/projectnetworktechnologies/infrastructure/service/UserService.java
@@ -1,8 +1,12 @@
 package com.borovyknt.projectnetworktechnologies.infrastructure.service;
 
+import com.borovyknt.projectnetworktechnologies.controller.dto.user.CreateUserDto;
+import com.borovyknt.projectnetworktechnologies.controller.dto.user.CreateUserResponseDto;
+import com.borovyknt.projectnetworktechnologies.controller.dto.user.GetUserDto;
 import com.borovyknt.projectnetworktechnologies.infrastructure.entity.UserEntity;
 import com.borovyknt.projectnetworktechnologies.infrastructure.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.autoconfigure.security.SecurityProperties;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -16,19 +20,53 @@ public class UserService {
         this.userRepository = userRepository;
     }
 
-    public UserEntity getOne(long id){
-        return userRepository.findById(id).orElseThrow(() -> new RuntimeException("User not found"));
+    public GetUserDto getOne(long id){
+        var userEntity = userRepository.findById(id).orElseThrow(() -> new RuntimeException("User not found"));
+        return new GetUserDto(
+                userEntity.getUserId(),
+                userEntity.getUsername(),
+                userEntity.getPassword(),
+                userEntity.getRole(),
+                userEntity.getEmail(),
+                userEntity.getName()
+        );
     }
 
-    public List<UserEntity> getAll(){
-        return userRepository.findAll();
+    public List<GetUserDto> getAll(){
+        var users = userRepository.findAll();
+        return users.stream().map((user) -> new GetUserDto(
+                user.getUserId(),
+                user.getUsername(),
+                user.getPassword(),
+                user.getRole(),
+                user.getEmail(),
+                user.getName()
+        )).collect(java.util.stream.Collectors.toList());
     }
 
-    public UserEntity create(UserEntity user){
-        return userRepository.save(user);
+    public CreateUserResponseDto create(CreateUserDto user){
+        var userEntity = new UserEntity();
+        userEntity.setUsername(user.getUsername());
+        userEntity.setPassword(user.getPassword());
+        userEntity.setRole(user.getRole());
+        userEntity.setEmail(user.getEmail());
+        userEntity.setName(user.getName());
+
+        var newUser = userRepository.save(userEntity);
+        return new CreateUserResponseDto(
+                newUser.getUserId(),
+                newUser.getUsername(),
+                newUser.getPassword(),
+                newUser.getRole(),
+                newUser.getEmail(),
+                newUser.getName()
+        );
     }
 
     public void delete(long id){
+        if(!userRepository.existsById(id)){
+            throw new RuntimeException();
+        }
         userRepository.deleteById(id);
     }
 }
